requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
comtypes==1.4.13
gTTS==2.5.4
idna==3.11
packaging==25.0
pillow==12.0.0
playsound==1.2.2
pypiwin32==223
PyQt6==6.10.0
PyQt6-Qt6==6.10.0
PyQt6_sip==13.10.2
pytesseract==0.3.13
pyttsx3==2.99
pywin32==311
requests==2.32.5
urllib3==2.5.0
